fixes issues with it not properly walking boxes
