+++ go.mod
@@ -2,13 +2,16 @@ module github.com/gobuffalo/packr
 
 require (
 	github.com/gobuffalo/envy v1.6.7
-	github.com/gobuffalo/packd v0.0.0-20181031195726-c82734870264
-	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/gobuffalo/genny v0.0.0-20181030163439-ed103521b8ec
+	github.com/gobuffalo/logger v0.0.0-20181027193913-9cf4dd0efe46
+	github.com/gobuffalo/packd v0.0.0-20181103221656-16c4ed88b296
+	github.com/karrick/godirwalk v1.7.5
 	github.com/markbates/oncer v0.0.0-20181014194634-05fccaae8fc4
+	github.com/markbates/safe v1.0.1
 	github.com/pkg/errors v0.8.0
+	github.com/sirupsen/logrus v1.2.0
 	github.com/spf13/cobra v0.0.3
-	github.com/spf13/pflag v1.0.3 // indirect
 	github.com/stretchr/testify v1.2.2
-	golang.org/x/net v0.0.0-20181102091132-c10e9556a7bc // indirect
 	golang.org/x/sync v0.0.0-20180314180146-1d60e4601c6f
+	golang.org/x/tools v0.0.0-20181102223251-96e9e165b75e
 )
